Add native lib location to library path
When the @native_library_dest_dir is set, the cookbook will extract the native library file also in this location (e.g. to be used by non-root users, in paths where they have access to). Therefore, when we are using it, we really should be including it in the library path set in the conf file.