Added chefspec unit tests
Because this cookbook mainly exposes an LWRP (so there's no recipe to test against), chefspec tests need to test against the test cookbook's recipes (which are used for the integration tests anyway).
I added a couple of recipes to test:
a. the generic installation
b. the native_library extraction feature (as I just noticed a bug with that)